pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "discussion-bots"
version = "0.1.0"
description = ""
authors = ["Oleksii Markov <alex@markovs.me>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.30.1"
pydantic = "^2.7.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-mockito = "^0.0.4"
flake8 = "^7.0.0"
flake8-black = "^0.3.6"
flake8-isort = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
